[flash_ctrl,dv] tb refactoring
- removed 'most' of redundent seqs

+++ hw/ip/flash_ctrl/dv/tests/flash_ctrl_base_test.sv
@@ -58,6 +58,10 @@ class flash_ctrl_base_test #(
     void'($value$plusargs("serr_pct=%0d", cfg.serr_pct));
     void'($value$plusargs("derr_pct=%0d", cfg.derr_pct));
     void'($value$plusargs("ierr_pct=%0d", cfg.ierr_pct));
+    void'($value$plusargs("otf_num_rw=%0d", cfg.otf_num_rw));
+    void'($value$plusargs("otf_num_hr=%0d", cfg.otf_num_hr));
+    void'($value$plusargs("otf_wr_pct=%0d", cfg.otf_wr_pct));
+    void'($value$plusargs("otf_rd_pct=%0d", cfg.otf_rd_pct));
   endfunction
 
   task run_phase(uvm_phase phase);
@@ -71,9 +75,11 @@ class flash_ctrl_base_test #(
 
     repeat(run_cnt) begin
       run_seq(test_seq_s, phase);
-      env.virtual_sequencer.stop_sequences();
-      `uvm_info("Test", $sformatf("TESTEND %0d",++dbg_run_cnt), UVM_MEDIUM)
-      foreach (env.m_tl_agents[i]) env.m_tl_agents[i].monitor.pending_a_req.delete();
+      if (run_cnt > 1) begin
+        env.virtual_sequencer.stop_sequences();
+        `uvm_info("Test", $sformatf("TESTEND %0d",++dbg_run_cnt), UVM_MEDIUM)
+        foreach (env.m_tl_agents[i]) env.m_tl_agents[i].monitor.pending_a_req.delete();
+      end
     end
     phase.drop_objection(this);
